+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- fix terminal spamming due to process abortion on Windows
+
 ## Released changes
 
 ### 0.4.0.2